createIndex and createApp in @misk/components
* Uniform create functions for generating the `<App>` and bootstrapping objects usually found in `index.tsx` for all Misk Web Tabs
* Eliminates ~100 lines of boilerplate from every Misk Web Tab
* `@misk/components 0.0.68`
* [CLOSES #563]
