+++ misk/web/@misk/components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@misk/components",
-  "version": "0.0.66",
+  "version": "0.0.68",
   "description": "Microservice Kontainer Common Components",
   "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
   "main": "lib/web/@misk/components/components.js",
