+++ misk/web/@misk/components/src/builders/createApp.tsx
@@ -1,0 +1,14 @@
+import { ConnectedRouter } from "connected-react-router"
+import { History } from "history"
+import * as React from "react"
+
+export const createApp = (
+  routes: JSX.Element
+) => {  
+  return (({ history } : { history: History }) => (
+      <ConnectedRouter history={history}>
+        {routes}
+      </ConnectedRouter>
+    )
+  )
+}
